# p2-2w2/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Laboratorio 3 - Examen Copa Americas 2024'","description":"'Template SpringBoot'","contact":{"name":"Hernán Jesús Morais","email":"hmorais@frc.utn.edu.ar"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Template SpringBoot'"}],"paths":{"/games/{id}/result":{"get":{"tags":["game-controller"],"operationId":"getGameResult","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GameResult"}}}}}},"put":{"tags":["game-controller"],"operationId":"putGameResult","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"local_goals","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"visitor_goals","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GameResult"}}}}}}},"/games/{id}/predictions":{"get":{"tags":["game-controller"],"operationId":"getPrediction","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PredictionResponseDto"}}}}}}},"put":{"tags":["game-controller"],"operationId":"putPrediction","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictionRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PredictionResponseDto"}}}}}}},"/users/sign-up":{"post":{"tags":["user-controller"],"operationId":"signUp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/users/sign-in":{"post":{"tags":["user-controller"],"operationId":"signIn","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/users":{"get":{"tags":["user-controller"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}}},"/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUsers_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/users/available-email":{"get":{"tags":["user-controller"],"operationId":"checkAvailableEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/scores":{"get":{"tags":["score-controller"],"operationId":"getScores","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScoreDto"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"pong","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/games":{"get":{"tags":["game-controller"],"operationId":"getGames","parameters":[{"name":"fase_game","in":"query","required":false,"schema":{"type":"string","enum":["GROUP_STAGE","QUARTER_FINALS","SEMI_FINALS","FINAL","THIRD_PLACE"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Game"}}}}}}}},"/games/{id}":{"get":{"tags":["game-controller"],"operationId":"getGame","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Game"}}}}}}},"/games/predictions":{"get":{"tags":["game-controller"],"operationId":"getPrediction_1","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"include_unpredicted","in":"query","required":true,"schema":{"type":"boolean"}},{"name":"fase_game","in":"query","required":false,"schema":{"type":"string","enum":["GROUP_STAGE","QUARTER_FINALS","SEMI_FINALS","FINAL","THIRD_PLACE"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PredictionResponseDto"}}}}}}}}},"components":{"schemas":{"Game":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"local":{"$ref":"#/components/schemas/Team"},"visitor":{"$ref":"#/components/schemas/Team"},"stadium":{"type":"string"},"date":{"type":"string","format":"date-time"},"fase_game":{"type":"string","enum":["GROUP_STAGE","QUARTER_FINALS","SEMI_FINALS","FINAL","THIRD_PLACE"]}}},"GameResult":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"game":{"$ref":"#/components/schemas/Game"},"result":{"type":"string","enum":["LOCAL_WIN","VISITOR_WIN","TIE"]},"local_goals":{"type":"integer","format":"int32"},"visitor_goals":{"type":"integer","format":"int32"}}},"Team":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"PredictionRequestDto":{"required":["local_goals","visitor_goals"],"type":"object","properties":{"local_goals":{"type":"integer","format":"int32"},"visitor_goals":{"type":"integer","format":"int32"}}},"PredictionResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserDto"},"game":{"$ref":"#/components/schemas/Game"},"result":{"type":"string","enum":["LOCAL_WIN","VISITOR_WIN","TIE"]},"local_goals":{"type":"integer","format":"int32"},"visitor_goals":{"type":"integer","format":"int32"}}},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"last_name":{"type":"string"}}},"SignUpDto":{"required":["email","last_name","name","password"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"last_name":{"type":"string"}}},"SignInDto":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"ScoreDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"},"points":{"type":"integer","format":"int32"}}}}}}