# p2-2w2/src/main/java/ar/edu/utn/frc/tup/lc/iii/dtos/ScoreDto.java
package ar.edu.utn.frc.tup.lc.iii.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ScoreDto {

    private UserDto user;
    private Integer points;
}