# p2-2w2/src/main/java/ar/edu/utn/frc/tup/lc/iii/dtos/UserDto.java
package ar.edu.utn.frc.tup.lc.iii.dtos;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UserDto {

    private Long id;
    private String name;

    @JsonProperty("last_name")
    private String lastName;
    private String email;
}