# p2-2w2/src/main/java/ar/edu/utn/frc/tup/lc/iii/models/Game.java
package ar.edu.utn.frc.tup.lc.iii.models;

import com.fasterxml.jackson.annotation.JsonFormat;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Game {

    private Long id;
    private Team local;
    private Team visitor;
    private String stadium;

    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy HH:mm:ss")
    private LocalDateTime date;

    @JsonProperty("fase_game")
    private FaseGame faseGame;
}